+++ src/server/src/model/subjects.js
@@ -1,14 +1,14 @@
 subjects = [
-    {id:1,name:"Artes"},
-    {id:2,name:"Biologia"},
-    {id:3,name:"Ciências"},
-    {id:4,name:"Educação física"},
-    {id:5,name:"Física"},
-    {id:6,name:"Geografia"},
-    {id:7,name:"História"},
-    {id:8,name:"Matemática"},
-    {id:9,name:"Português"},
-    {id:10,name:"Química"},
+    {id:0,name:"Artes"},
+    {id:1,name:"Biologia"},
+    {id:2,name:"Ciências"},
+    {id:3,name:"Educação física"},
+    {id:4,name:"Física"},
+    {id:5,name:"Geografia"},
+    {id:6,name:"História"},
+    {id:7,name:"Matemática"},
+    {id:8,name:"Português"},
+    {id:9,name:"Química"},
 ]
 
 module.exports = subjects