+++ src/server/src/utils.js
@@ -1,0 +1,14 @@
+function getSubject(subjectNumber) {
+    return subjects[subjectNumber];
+  }
+
+
+  function convertHoursToMinutes(time){
+    let [hours,minutes] = time.split(":")
+    return Number((hours*60) + minutes)
+  }
+
+module.exports = {
+    getSubject,
+    convertHoursToMinutes
+}